#1288 Started to add tests for new extraction driver creation.
