+++ src/psyclone/tests/domain/gocean/transformations/gocean_extract_test.py
@@ -584,181 +584,6 @@ def test_driver_generation_flag(tmpdir, create_driver):
 
 
 # -----------------------------------------------------------------------------
-def test_driver_creation(tmpdir):
-    '''Test that driver is created correctly for all variable access \
-    modes (input, input-output, output).
-
-    '''
-    # Use tmpdir so that the driver is created in tmp
-    tmpdir.chdir()
-
-    etrans = GOceanExtractTrans()
-    ctrans = GOConstLoopBoundsTrans()
-    psy, invoke = get_invoke("driver_test.f90",
-                             GOCEAN_API, idx=0, dist_mem=False)
-    schedule = invoke.schedule
-    # This test expects constant loop bounds
-    ctrans.apply(schedule)
-
-    etrans.apply(schedule.children[0], {'create_driver': True})
-    # We are only interested in the driver, so ignore results.
-    str(psy.gen)
-
-    driver = tmpdir.join("driver-psy_extract_example_with_various_variable_"
-                         "access_patterns-invoke_0_compute_kernel:compute_"
-                         "kernel_code:r0.f90")
-    assert driver.isfile()
-
-    with driver.open("r") as driver_file:
-        driver_code = driver_file.read()
-
-    # This is an excerpt of the code that should get created.
-    # It is tested line by line since there is other code in between
-    # which is not important, and the order might also change. It also
-    # tests if unique variable names are created in the driver: the user
-    # program contains a local variable 'dx', which clashes with the grid
-    # property dx. The grid property will be renamed to 'dx_1':
-    expected = '''use extract_psy_data_mod, only : extract_PsyDataType
-
-  real*8, allocatable, dimension(:,:) :: out_fld
-  real*8, allocatable, dimension(:,:) :: in_out_fld
-  real*8, allocatable, dimension(:,:) :: in_fld
-  real*8, allocatable, dimension(:,:) :: dx
-  real*8, allocatable, dimension(:,:) :: in_fld_grid_gphiu
-  real*8, allocatable, dimension(:,:) :: out_fld_post
-  real*8 :: in_fld_grid_dx
-  real*8, allocatable, dimension(:,:) :: in_out_fld_post
-  type(extract_PsyDataType) :: extract_psy_data
-  call extract_psy_data%OpenRead('psy_extract_example_with_various_variable_''' \
-  '''access_patterns', 'invoke_0_compute_kernel:compute_kernel_code:r0')
-  call extract_psy_data%ReadVariable('out_fld_post', out_fld_post)
-  ALLOCATE(out_fld(SIZE(out_fld_post, 1), SIZE(out_fld_post, 2)))
-  out_fld = 0
-  call extract_psy_data%ReadVariable('in_fld', in_fld)
-  call extract_psy_data%ReadVariable('in_out_fld_post', in_out_fld_post)
-  call extract_psy_data%ReadVariable('dx', dx)
-  call extract_psy_data%ReadVariable('in_fld%grid%dx', in_fld_grid_dx)
-  call extract_psy_data%ReadVariable('in_fld%grid%gphiu', in_fld_grid_gphiu)
-  do j = 2, jstop, 1
-    do i = 2, istop+1, 1
-      call compute_kernel_code(i, j, out_fld, in_out_fld, in_fld, dx, ''' \
-      '''in_fld_grid_dx, in_fld_grid_gphiu)
-    enddo
-  enddo
-  if (i == i_post) then
-    PRINT *, "i correct"
-  else
-    PRINT *, "i incorrect. Values are:"
-    PRINT *, i
-    PRINT *, "i values should be:"
-    PRINT *, i_post
-  end if
-  if (ALL(in_out_fld - in_out_fld_post == 0.0)) then
-    PRINT *, "in_out_fld correct"
-  else
-    PRINT *, "in_out_fld incorrect. Values are:"
-    PRINT *, in_out_fld
-    PRINT *, "in_out_fld values should be:"
-    PRINT *, in_out_fld_post
-  end if'''
-    expected_lines = expected.split("\n")
-    for line in expected_lines:
-        assert line in driver_code
-
-
-# -----------------------------------------------------------------------------
-def test_rename_suffix_if_name_clash(tmpdir):
-    '''Test that driver is created correctly if there is a clash
-    with the variable names, e.g. an output variable 'a', and
-    an input variable 'a_post' - writing the output variable 'a'
-    would use 'a_post' as name, so the suffix must be changed.
-
-    '''
-    # Use tmpdir so that the driver is created in tmp
-    tmpdir.chdir()
-
-    etrans = GOceanExtractTrans()
-    psy, invoke = get_invoke("driver_test.f90",
-                             GOCEAN_API, idx=1, dist_mem=False)
-    schedule = invoke.schedule
-
-    etrans.apply(schedule.children[0], {'create_driver': True})
-    extract_code = str(psy.gen)
-
-    # Due to the name clash of "out_fld"+"_post" and "out_fld_post"
-    # the _post suffix is changed to _post0. So the file will
-    # contain out_fld_post for the input variable out_fld_post,
-    # and "out_fld_post0" for the output value of out_fld.
-    expected = """
-      CALL extract_psy_data%PreDeclareVariable("out_fld_post", out_fld_post)
-      CALL extract_psy_data%PreDeclareVariable("in_out_fld_post0", in_out_fld)
-      CALL extract_psy_data%PreDeclareVariable("out_fld_post0", out_fld)
-      CALL extract_psy_data%ProvideVariable("in_out_fld", in_out_fld)
-      CALL extract_psy_data%ProvideVariable("out_fld_post", out_fld_post)
-      CALL extract_psy_data%ProvideVariable("in_out_fld_post0", in_out_fld)
-      CALL extract_psy_data%ProvideVariable("out_fld_post0", out_fld)"""
-    expected_lines = expected.split("\n")
-    ordered_lines_in_text(expected_lines, extract_code)
-
-    # Now we also need to check that the driver uses the new suffix,
-    # i.e. both as key for ReadVariable, as well as for the variable
-    # names.
-    driver = tmpdir.join("driver-psy_extract_example_with_various_variable_"
-                         "access_patterns-invoke_1_compute_kernel:compute_"
-                         "kernel_code:r0.f90")
-    assert driver.isfile()
-
-    with driver.open("r") as driver_file:
-        driver_code = driver_file.read()
-
-    expected = """
-  real*8, allocatable, dimension(:,:) :: out_fld
-  real*8, allocatable, dimension(:,:) :: in_out_fld
-  real*8, allocatable, dimension(:,:) :: out_fld_post
-  real*8, allocatable, dimension(:,:) :: in_out_fld_post0
-  real*8, allocatable, dimension(:,:) :: out_fld_post0
-  call extract_psy_data%ReadVariable('in_out_fld', in_out_fld)
-  call extract_psy_data%ReadVariable('in_out_fld_post0', in_out_fld_post0)
-  call extract_psy_data%ReadVariable('out_fld_post0', out_fld_post0)
-  ALLOCATE(out_fld(SIZE(out_fld_post0, 1), SIZE(out_fld_post0, 2)))
-  call extract_psy_data%ReadVariable('out_fld_post', out_fld_post)"""
-
-    ordered_lines_in_text(expected.split("\n"), driver_code)
-
-    # Now test that more than one variable clash is handled. The third
-    # invoke uses:
-    # "out_fld" as output field
-    # "out_fld_post" as input field (first clash --> suffix becomes "_post0")
-    # "out_fld_post0" as input+output field (next clash --> suffix = "_post1")
-    psy, invoke = get_invoke("driver_test.f90",
-                             GOCEAN_API, idx=2, dist_mem=False)
-    schedule = invoke.schedule
-    # We don't check the driver, we already tested that the
-    # driver picks up the adjusted suffix above
-    etrans.apply(schedule.children[0])
-    extract_code = str(psy.gen)
-
-    # Check that *out_fld* is declared correctly: it is only declared as
-    # output value, so must use key out_fld_post1 once, and not be declared
-    # as input value:
-    assert 'PreDeclareVariable("out_fld_post1", out_fld)' in extract_code
-    assert 'PreDeclareVariable("out_fld", out_fld)' not in extract_code
-
-    # Check that *out_fld_post* (input/output) is declared correctly. It must
-    # be declared twice: once for the input value using the original variable
-    # name, and once as output using the "_post1" suffix"
-    assert 'PreDeclareVariable("out_fld_post", out_fld_post)' in extract_code
-    assert 'PreDeclareVariable("out_fld_post_post1", out_fld_post)' \
-        in extract_code
-
-    # Check that *out_fld_post0* is declared correctly: as input-only
-    # variable it must be declared once for using the original variable name.
-    assert 'PreDeclareVariable("out_fld_post0", out_fld_post0)' in extract_code
-    assert 'PreDeclareVariable("out_fld_post0_post1", out_fld_post0)' \
-        not in extract_code
-
-
-# -----------------------------------------------------------------------------
 def test_driver_loop_variables(tmpdir):
     '''Test that loop variables are not stored. ATM this test
     fails because of #641.
